# public/openapi/spec.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Trukt
  description: Transportation management system
servers:
  - url: 'https://truktbackend-dev.rcgauto.p.azurewebsites.net'
    description: Development server
  - url: 'https://truktbackend-staging.rcgauto.p.azurewebsites.net'
    description: Staging server
  - url: 'https://truktbackend.rcgauto.p.azurewebsites.net'
    description: Production server
paths:
...
  '/statusmanager/{guid}':
    get:
      summary: Get status log by order guid
      operationId: getStatusLog
      description: Get status log by order guid
      tags:
        - Statusmanager
      parameters:
        - in: path
          name: guid
          schema:
            $ref: '#/components/schemas/guid'
          required: true
        - in: query
          name: pg
          schema:
            $ref: '#/components/schemas/page'
        - in: query
          name: rc
          schema:
            $ref: '#/components/schemas/rowCount'
      responses:
        '200':
          description: status log
          content:
            application/json: {}
...
components:
  schemas:
    guid:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-(?:[0-9a-f]{4}-){3}[0-9a-f]{12}$'
      example: 00000000-0000-0000-0000-000000000000
...
    page:
      description: 'For pagination purposes, page to return.'
      type: integer
      minimum: 1
      default: 1
      example: 1
    rowCount:
      description: 'For pagination purposes, number of entries per page.'
      type: integer
      default: 25
...